requirements.txt
faiss-cpu
sentence-transformers
openai
python-dotenv
